requirements.txt
aiogram==2.20
PyYAML==5.3.1